requirements.txt
langchain-core>=0.2.30
langchain-community>=0.2.16
langchain-google-genai>=2.0.4
langchain-text-splitters>=0.2.4
faiss-cpu>=1.8.0.post1
fastapi>=0.114.0
uvicorn>=0.30.6
pypdf>=4.3.1
docx2txt>=0.8
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
neonize>=0.3.14.post0
